Email functionality edited & JavaDoc comments inserted
Email feature: now the “from” email address is the user’s default email
address. The “to” field is populated by “example@gmail.com” indicating
to the user that the recipient must be entered here.

JavaDoc comments written for each function in MainActivity, to provide
anyone reading the code with a context for what each function does.