Set-Up Glitch Solution but not implemented
Note that currently the Floating Action Menu Button is not at the
bottom right where it should be. This is because a ScrollView must have
its height set to “wrap_content” rather than “match_parent”.

When the floating action menu was placed in the activity_main file, the
desired result was achieved, however for some reason the click
functions of the drop down buttons were disabled.

Hence the floating action menu implementation in activity_main has been
commented out for now, and is still in content_main, under the
ScrollView.

**SOLUTION PENDING**