Changes to NotificationHandler, MakeNotification, and Main Activity
Notification Handler
TimerTask added in order to run the createNotification function every 24
hours

MakeNotification
Calendar parameter added in order to take in the current time and
compare the current date with the date of the task.

Main Activity
For loop added in the actionAdd onClickListener. This passes the date
and description of each task in the database to into the
showNotification function.